requirements.txt
pandas==2.2.2
tqdm==4.66.5
pdfmerge==1.0.0
openpyxl==3.1.5
numpy==2.0.1
pillow==10.4.0